+++ src/Command/CommandInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smeghead\PhpVariableHardUsage\Command;
+
+interface CommandInterface
+{
+    public function execute(): void;
+}